Improve whitespace support in wiki pages
It's quite common for wiki pages to contain whitespaces, since that
makes for a much nicer page title and navigation tree in AzDO.

These URIs would previously be broken in that the whitespace would
not be URL-encoded and pasting it in slack or email would result in
a broken experience.

I took the opportunity to implement an improvement that makes the
URLs even better in this very common scenario, by replacing the
spaces with dashes instead by default. Obviously, dashes are valid
in wiki page names and folders, so if we find existing dashes, we
convert spaces to underscores instead. Unless there are also underscores
and in that case we just URL encode the whole thing.

When converting using underscores or not converting at all, we hint
the azure function by appending a `?u` or `?b` (for *u*nderscores or
*b*are) so that the function knows how to convert back to the final
redirect location.